Moving fit task to server dir.
